% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_outcome_scb.R
\name{check_subset}
\alias{check_subset}
\title{Check validity of subset specification against a data set}
\usage{
check_subset(subset, data_fit)
}
\arguments{
\item{subset}{A \code{data.frame} or a named \code{list} specifying filtering conditions.
\itemize{
\item If \code{data.frame}, column names are treated as variable names, and column values are the allowed values.
\item If \code{list}, names are variable names, and elements are the allowed values.
Each variable name must exist in \code{data_fit}.
}}

\item{data_fit}{A \code{data.frame} containing the data to be checked against.
All variable names and values specified in \code{subset} must appear in this dataset.}
}
\value{
Invisibly returns \code{TRUE} if all checks pass; otherwise stops with an error message.
}
\description{
This function verifies that the user-supplied \code{subset} is valid.
\code{subset} must be either a \code{data.frame} or a named \code{list}.
Each element (or column) corresponds to a variable in \code{data_fit},
and its values must exist among the observed values or levels of the
corresponding variable in \code{data_fit}.
}
\examples{
# Used internally by SCB_gls_climate

}
\keyword{internal}
