% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_outcome_scb.R
\name{expit}
\alias{expit}
\title{Expit (inverse logit) function}
\usage{
expit(x)
}
\arguments{
\item{x}{A numeric input.}
}
\value{
Value between 0 and 1.
}
\description{
Computes the inverse logit transformation.
}
\examples{
expit(0)         # returns 0.5
expit(c(-2, 0, 2))
}
