% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scb_to_cs.R
\name{incl_f}
\alias{incl_f}
\title{Test inclusion relationship between two logical vectors}
\usage{
incl_f(A, B)
}
\arguments{
\item{A}{A logical vector.}

\item{B}{A logical vector of the same length as \code{A}.}
}
\value{
A logical value: \code{TRUE} if all \code{TRUE} values in \code{A} are also \code{TRUE} in \code{B}; otherwise \code{FALSE}.
}
\description{
This function tests whether all \code{TRUE} positions in vector \code{A} are also \code{TRUE} in vector \code{B},
which is equivalent to testing whether \code{A} is element-wise included in \code{B}.
}
\examples{
# Used internally by scb_to_cs

}
\keyword{internal}
