% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invSCI-package.R
\docType{package}
\name{invSCI-package}
\alias{invSCI}
\alias{invSCI-package}
\title{invSCI: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
