% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocph.R
\name{rocph}
\alias{rocph}
\title{rocph
Tracé d'une courbe ROC simple avec l'AUC}
\usage{
rocph(zz, varx, titre, cutx)
}
\arguments{
\item{zz}{valeur numerique a tester}

\item{varx}{variable de test, binaire}

\item{titre}{titre}

\item{cutx}{Valeurs de zz a inscrire sur la courbe}
}
\value{
Courbe ROC
}
\description{
rocph
Tracé d'une courbe ROC simple avec l'AUC
}
\examples{
a=1
}
