% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_dense.R
\name{SCB_dense}
\alias{SCB_dense}
\title{Construct Simultaneous Confidence Bands (SCB) for Dense Functional Data}
\usage{
SCB_dense(A, mean_A = NULL, alpha_level = 0.05)
}
\arguments{
\item{A}{An array of dimension (D₁, D₂, ..., N), where N is the number of repetitions.}

\item{mean_A}{Optional array of same shape as \code{A[,,1]}, representing the mean function.}

\item{alpha_level}{Significance level for SCB. Default is 0.05.}
}
\value{
A list with upper and lower SCB bounds and the threshold used.
}
\description{
Construct Simultaneous Confidence Bands (SCB) for Dense Functional Data
}
\examples{
A <- array(rnorm(100 * 50), dim = c(100, 50))
result <- SCB_dense(A)

data(ccds_fpca)
data(ccds_cma)

Y_samples <- ccds_fpca$percent_change
n_time <- length(unique(ccds_fpca$seconds))
n_subject <- length(Y_samples) / n_time
# transform to matrix (T, N)
Y_mat <- matrix(Y_samples, nrow = n_time, ncol = n_subject)
Y_mat[is.na(Y_mat)] <- 0 # replace NA
SCB_dense(Y_mat, ccds_cma$yhat)

}
