% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valscore.R
\name{valscore}
\alias{valscore}
\title{Courbe ROC pour un score}
\usage{
valscore(sc, varx, titre = "Score")
}
\arguments{
\item{sc}{: score à tester}

\item{varx}{: variable d'outcome}

\item{titre}{: titre du graphique}
}
\value{
Courbe ROC
}
\description{
Courbe ROC pour un score
}
\examples{
a=1
}
