% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_ccds.R
\docType{data}
\name{ccds}
\alias{ccds}
\title{Load Raw ccds longitudinal data from clinical study and Clean the dataset (Right Eye, Post Timepoint)}
\format{
A tibble with N rows and 4 variables:
\describe{
\item{subject_id}{Character. Subject identifier.}
\item{seconds}{Numeric. Time in seconds since initial measurement.}
\item{use}{Numeric. Binary indicator of usage status; \code{1} indicates "use", \code{0} otherwise.}
\item{percent_change}{Numeric. Percent change in the outcome of interest.}
\item{subject}{Character. Equal to subject_id.}
}
}
\source{
Processed from \code{data-raw/ccds_load.R} using the \code{here} and \code{dplyr} packages.
}
\usage{
data(ccds)
}
\description{
This dataset contains repeated measures of percent change over time
for multiple subjects under two use categories. It includes both user and non-user groups,
time points, and metadata related to eye side and frame timing. cleaning process make sure that the data only includes measurements taken from the right eye
at the post-intervention timepoint (\code{tp == "post"}).
}
\details{
The data was filtered to include only right-eye measurements at the post timepoint from the original \code{ccds_raw} dataset.
The \code{use} variable is derived from a character column, with \code{"use"} mapped to \code{1}, and all other values to \code{0}.
}
\keyword{dataset}
