% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\name{cma}
\alias{cma}
\title{Construct CMA Confidence Intervals via Parametric Method}
\usage{
cma(
  object,
  model,
  time,
  range = NULL,
  groups = NULL,
  subject = NULL,
  nboot = NULL
)
}
\arguments{
\item{object}{A fitted functional outcome regression model object (e.g., from mgcv::bam()/mgcv::gam()).}

\item{model}{A character string specifying the model type used (e.g., gam or bam).}

\item{time}{A character string specifying the name of the time variable used in the model.}

\item{range}{A numeric vector specifying the range or grid of time points interested. Defauts is the whole time points from the fitted model.}

\item{groups}{A character vector specifying the names of group-related covariates of interest. If NULL, CIs are for the reference group.}

\item{subject}{A character string specifying the name of the subject-level random effect variable.}

\item{nboot}{An integer specifying the number of bootstrap samples used to construct the confidence bands. Defaults to 10,000.}
}
\value{
A list containing:
\item{yhat}{Estimated mean function for the group of interest.}
\item{time}{The time points used.}
\item{se_hat}{Standard errors of the estimated means.}
\item{scb_low}{Lower bound of the simultaneous confidence band.}
\item{scb_uo}{Upper bound of the simultaneous confidence band.}
\item{type}{A character description of the output type.}
}
\description{
This function computes Correlation and Multiplicity Adjusted (CMA) confidence bands for a specified group in a functional outcome regression model
using parameter simulations approach with Gaussian multiplier bootstrap.
}
\examples{
# example using ccds data
data(ccds_fpca)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(subject, by = Phi1, bs="re") +
  s(subject, by = Phi2, bs="re")+
  s(subject, by = Phi3, bs="re") +
  s(subject, by = Phi4, bs="re"),
  method = "fREML", data = ccds_fpca, discrete = TRUE)
cma(fosr_mod, "gam", "seconds", groups = "use", subject = "subject")

}
