% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\name{cma}
\alias{cma}
\title{Construct CMA Confidence Intervals via Parametric Method}
\usage{
cma(
  data_df,
  object,
  fitted = TRUE,
  alpha = 0.05,
  time,
  range = NULL,
  group_name = NULL,
  group_value = NULL,
  subject = NULL,
  nboot = NULL
)
}
\arguments{
\item{data_df}{A functional data frame.
The input data must have column names, and should contain the functional outcome, time and subject}

\item{object}{A fitted Function-on-Scalar Regression (FoSR) object (e.g., from mgcv::bam()/mgcv::gam()).}

\item{fitted}{Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function
\itemize{
\item \code{TRUE} - Estimate the simultaneous confidence bands for fitted mean outcome function.
\item \code{FALSE} - estimate the simultaneous confidence bands for fitted parameter function.
}
Default is \code{TRUE}.}

\item{alpha}{Significance level for SCB. Default is 0.05.}

\item{time}{A character string specifying the name of the time variable used in the model.}

\item{range}{A numeric vector specifying the range or grid of time points interested. Defaut is the whole time points from the fitted model.}

\item{group_name}{A character vector that specifies the names of grouping scalar variables to analyze.
Default is \code{NULL}, representing the reference group.}

\item{group_value}{A numeric vector that specifies the corresponding values of the variables listed in \code{group_name}.
Each entry in \code{group_value} should match the respective variable in \code{group_name}.
Character/factor is not allowed.
Default is \code{NULL}, representing the reference group.}

\item{subject}{A optional character string specifying the name of the subject-level random effect variable, if included in model fitting.}

\item{nboot}{An integer specifying the number of bootstrap samples used to construct the confidence bands. Default is 10,000.}
}
\value{
A list containing:
\item{yhat}{Estimated mean function for the group of interest.}
\item{time}{The time points used.}
\item{se_hat}{Standard errors of the estimated means.}
\item{scb_low}{Lower bound of the simultaneous confidence band.}
\item{scb_up}{Upper bound of the simultaneous confidence band.}
\item{type}{A character description of the output type.}
}
\description{
This function computes Correlation and Multiplicity Adjusted (CMA) confidence bands for a specified group in a functional outcome regression model
using parameter simulations approach with Gaussian multiplier bootstrap.
}
\examples{
# example using pupil data
data(pupil)
pupil_fpca <- prepare_pupil_fpca(pupil)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(id, by = Phi1, bs="re") +
  s(id, by = Phi2, bs="re")+
  s(id, by = Phi3, bs="re") +
  s(id, by = Phi4, bs="re"),
  method = "fREML", data = pupil_fpca, discrete = TRUE)

results <- cma(pupil_fpca, fosr_mod, fitted = TRUE, time = "seconds",
group_name = "use", group_value = 1, subject = "id")

results <- cma(pupil_fpca, fosr_mod, fitted = FALSE, time = "seconds",
group_name = "use", group_value = 1, subject = "id")

}
