% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\name{mean_response_predict}
\alias{mean_response_predict}
\title{Functional Outcome Regression Prediction with Group-Specific Inference}
\usage{
mean_response_predict(
  data_df,
  object,
  fitted = TRUE,
  time,
  range = NULL,
  group_name,
  group_value,
  subject = NULL
)
}
\arguments{
\item{data_df}{A functional data frame.
The input data must have column names, and should contain the functional outcome, time and subject}

\item{object}{A fitted Function-on-Scalar Regression (FoSR) model object (e.g., from mgcv::bam()/mgcv::gam()).}

\item{fitted}{Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function
\itemize{
\item \code{TRUE} - Estimate the simultaneous confidence bands for fitted mean outcome function.
\item \code{FALSE} - estimate the simultaneous confidence bands for fitted parameter function.
}
Default is \code{TRUE}.}

\item{time}{A character string specifying the name of the time variable used in the model.}

\item{range}{A numeric vector specifying the range or grid of time points interested. Defaut is the whole time points from the fitted model.}

\item{group_name}{A character vector that specifies the names of grouping scalar variables to analyze.
Default is \code{NULL}, representing the reference group.}

\item{group_value}{A numeric vector that specifies the corresponding values of the variables listed in \code{group_name}.
Each entry in \code{group_value} should match the respective variable in \code{group_name}.
Character/factor is not allowed.
Default is \code{NULL}, representing the reference group.}

\item{subject}{A optional character string specifying the name of the subject-level random effect variable, if included in model fitting.}
}
\value{
A list containing the following elements:
\describe{
\item{s_pred}{Numeric vector of sorted unique time points used for prediction}
\item{pred_df}{Data frame with prediction results, containing:
\itemize{
\item \code{mean}: Predicted mean values
\item \code{se}: Standard errors
}
}
\item{lpmat}{Linear predictor matrix (design matrix) used for confidence interval calculations}
\item{mod_coef}{Vector of model coefficients for selected group}
\item{mod_cov}{Variance-covariance matrix corresponding to the selected group coefficients}
}
}
\description{
This function is an internal function for constructing SCBs for functional data.
}
\examples{
data(ccds)
ccds_fpca <- prepare_ccds_fpca(ccds)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(subject, by = Phi1, bs="re") +
  s(subject, by = Phi2, bs="re")+
  s(subject, by = Phi3, bs="re") +
  s(subject, by = Phi4, bs="re"),
  method = "fREML", data = ccds_fpca, discrete = TRUE)

results <- mean_response_predict(ccds_fpca, fosr_mod, fitted = TRUE, time = "seconds",
group_name = "use", group_value = 1, subject = "subject")

results <- mean_response_predict(ccds_fpca, fosr_mod, fitted = FALSE, time = "seconds",
group_name = "use", group_value = 1, subject = "subject")

}
