% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cma.R
\name{prepare_pupil_fpca}
\alias{prepare_pupil_fpca}
\title{Prepare pupil FPCA Dataset}
\usage{
prepare_pupil_fpca(input_data, k_mean = 30, k_fpca = 15)
}
\arguments{
\item{input_data}{Raw ccds data frame containing:
\itemize{
\item \code{percent_change}: Functional response variable
\item \code{seconds}: Time variable
\item{smoker}{Numeric. 1 = Daily or Occasional user, 0 = No Use.}
\item{Daily}{Numeric. 1 = Daily user only, 0 = otherwise.}
\item{gender}{Numeric. Binary gender: 1 = Female, 0 = Male.}
}}

\item{k_mean}{Number of basis functions for mean model smooth terms (default: 30)}

\item{k_fpca}{Number of knots for FPCA estimation (default: 15)}
}
\value{
A tibble containing:
\itemize{
\item Original ccds variables
\item FPCA eigenfunctions (Phi1, Phi2,...)
\item Sorted by subject and time
}
}
\description{
Processes data by fitting a mean GAM model, extracting residuals, performing FPCA,
and merging the results to create an enhanced dataset for functional regression analysis.
}
\examples{
data(pupil)
processed_data <- prepare_pupil_fpca(pupil)

}
