% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCB_functional_outcome.R
\name{SCB_functional_outcome}
\alias{SCB_functional_outcome}
\title{Construct Simultaneous Confidence Bands (SCB) For One Dimensional Functional Data Using Specified Methods}
\usage{
SCB_functional_outcome(
  data_df,
  object = NULL,
  method,
  fitted = TRUE,
  est_mean = TRUE,
  alpha = 0.05,
  outcome,
  time,
  range = NULL,
  group_name = NULL,
  group_value = NULL,
  subject,
  nboot = NULL,
  method_SD = "t",
  weights = "rademacher"
)
}
\arguments{
\item{data_df}{A functional data frame.
The input data must have column names, and should contain the functional outcome, time, subject}

\item{object}{A fitted Function-on-Scalar Regression (FoSR) object (e.g., from mgcv::bam()/mgcv::gam()). Default is \code{NULL}}

\item{method}{A character string specifying the approach:
\itemize{
\item \code{"cma"} - Correlation and Multiplicity Adjusted (CMA) confidence bands
via parametric approach (requires a fitted functional regression model)
\item \code{"multiplier"} - Dense confidence bands via Multiplier-t Bootstrap method
For \code{method = "multiplier"} option, the user should check whether their input data (\code{data}) has zero entries on every subject for a specific time point.
The function only allow all zero entries for time point zero. Otherwise, the function will return an error.
Besides, for Multiplier Bootstrap, if there exist NA in data, the function will impute these NA's via fpca.face.
}}

\item{fitted}{Logical. Whether to estimate the simultaneous confidence bands for fitted mean function or fitted parameter function
\itemize{
\item \code{TRUE} - Estimate the simultaneous confidence bands for fitted mean outcome function.
\item \code{FALSE} - estimate the simultaneous confidence bands for fitted parameter function.
}
Default is \code{TRUE}.}

\item{est_mean}{Logical. Whether to use the fitted mean from a functional regression model in Multiplier Bootstrap (\code{method = "multiplier"}) for estimating the quantile.
This argument can be ignored if choose \code{method = "cma"}.
If \code{TRUE}, the function will use:
\itemize{
\item \strong{Fitted mean:} The estimated mean function from a fitted model (e.g., \code{mgcv::bam}).
Need to provide a fitted functional outcome regression model object. Otherwise, the sample mean will be calculated.
}
If \code{FALSE}, the function will instead use:
\itemize{
\item \strong{Sample mean:}
\deqn{\hat{\beta}(s) = \frac{1}{n} \sum_{i=1}^n \tilde{Y}_i(s)}
where \eqn{\tilde{Y}_i(s)} is a smoothed version of the observed functional response.
For this sample mean option, if there exists NA in \code{data}, will perform imputation to fill in those values.
}
Default is \code{FALSE}.}

\item{alpha}{Significance level for SCB. Default is 0.05.}

\item{outcome}{A character string specifying the name of the outcome variable used in the model. Must be provided if choose \code{method = "multiplier"}.}

\item{time}{A character string specifying the name of the time variable used in the model.}

\item{range}{A numeric vector specifying the range or grid of time points interested. Defaut is the whole time points from the fitted model.}

\item{group_name}{A character vector that specifies the names of grouping scalar variables to analyze.
Default is \code{NULL}, representing the reference group if \code{method = "cma"}.
For \code{method = "multiplier"}, \code{NuULL} is not allowed.}

\item{group_value}{A numeric vector that specifies the corresponding values of the variables listed in \code{group_name}.
Each entry in \code{group_value} should match the respective variable in \code{group_name}.
Character/factor is not allowed.
Default is \code{NULL}, representing the reference group if \code{"cma"}.
For \code{method = "multiplier"}, \code{NuULL} is not allowed.}

\item{subject}{A character string specifying the name of the subject variable.}

\item{nboot}{An integer specifying the number of bootstrap samples used to construct the confidence bands. Default is 10,000 for cma, 5000 for Multiplier Bootstrap.}

\item{method_SD}{Method for SD estimation: "t" or "regular". Default is "t".}

\item{weights}{Multiplier type: "rademacher", "gaussian", or "mammen". Default is "rademacher"."

For CMA bands (\code{method = "cma"}), please provide a pre-fitted functional regression model
For Multiplier Bootstrap bands (\code{method = "multiplier"}), the original input data is required.}
}
\value{
A list containing:
\item{yhat}{Estimated mean function for the group of interest.}
\item{time}{The time points used.}
\item{se_hat}{Standard errors of the estimated means.}
\item{scb_low}{Lower bound of the simultaneous confidence band.}
\item{scb_up}{Upper bound of the simultaneous confidence band.}
\item{type}{A character description of the output type.}
}
\description{
This function builds simultaneous confidence bands through two distinct approaches..
}
\examples{
# example using pupil data
data(pupil)
pupil_fpca <- prepare_pupil_fpca(pupil)

fosr_mod <- mgcv::bam(percent_change ~ s(seconds, k=30, bs="cr") +
  s(seconds, by = use, k=30, bs = "cr") +
  s(id, by = Phi1, bs="re") +
  s(id, by = Phi2, bs="re")+
  s(id, by = Phi3, bs="re") +
  s(id, by = Phi4, bs="re"),
  method = "fREML", data = pupil_fpca, discrete = TRUE)

# CMA approach
results <- SCB_functional_outcome(data_df = pupil, object = fosr_mod, method = "cma", fitted = TRUE,
                                  est_mean = TRUE, outcome = "percent_change", time = "seconds",
                                  group_name = "use", group_value = 1, subject = "id")


# multiplier bootstrap
results <- SCB_functional_outcome(data_df = pupil, object = fosr_mod, method = "multiplier", fitted = TRUE,
                       est_mean = TRUE, outcome = "percent_change",
                       time = "seconds", group_name = "use", group_value = 1, subject = "id")
results <- SCB_functional_outcome(data_df = pupil, object = fosr_mod, method = "multiplier", fitted = TRUE,
                       est_mean = FALSE, outcome = "percent_change",
                       time = "seconds", group_name = "use", group_value = 1, subject = "id")


}
