% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCoRES-package.R
\docType{package}
\name{SCoRES-package}
\alias{SCoRES}
\alias{SCoRES-package}
\title{SCoRES: Simultaneous Confidence Region Excursion Sets}
\description{
Provides computational tools for estimating inverse regions and constructing the corresponding simultaneous outer and inner confidence regions. Acceptable input includes both one-dimensional and two-dimensional data for linear, logistic, functional, and spatial generalized least squares regression models. Functions are also available for constructing simultaneous confidence bands (SCBs) for these models. The definition of simultaneous confidence regions (SCRs) follows Sommerfeld et al. (2018) \doi{10.1080/01621459.2017.1341838}. Methods for estimating inverse regions, SCRs, and the nonparametric bootstrap are based on Ren et al. (2024) \doi{10.1093/jrsssc/qlae027}. Methods for constructing SCBs are described in Crainiceanu et al. (2024) \doi{10.1201/9781003278726} and Telschow et al. (2022) \doi{10.1016/j.jspi.2021.05.008}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://angelayustat.github.io/SCoRES/}
  \item Report bugs at \url{https://github.com/AngelaYuStat/SCoRES/issues}
}

}
\author{
\strong{Maintainer}: Zhuoran Yu \email{angela.yu@emory.edu}

Authors:
\itemize{
  \item Armin Schwartzman \email{armins@health.ucsd.edu}
  \item Junting Ren \email{junting.ren.stat@gmail.com}
  \item Julia Wrobel \email{julia.wrobel@emory.edu}
}

}
\keyword{internal}
