% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SCoRES-package.R
\docType{package}
\name{SCoRES-package}
\alias{SCoRES}
\alias{SCoRES-package}
\title{SCoRES: Simultaneous COnfidence Region Excursion Sets}
\description{
This package provides useful statistical tools for both the estimation of the inverse set and the corresponding simultaneous outer and inner confidence sets. Acceptable forms of input includes both 1D and 2D data for linear regression, logistic regression, functional regression and spatial generalized least square regression. Useful functions are also provided for constructing simultaneous confidence bands (SCB) for these models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://angelayustat.github.io/SCoRES}
  \item Report bugs at \url{https://github.com/AngelaYuStat/SCoRES/issues}
}

}
\author{
\strong{Maintainer}: Angela Yu \email{angela.yu@emory.edu}

Authors:
\itemize{
  \item Armin Schwartzman \email{armins@health.ucsd.edu}
  \item Julia Wrobel \email{julia.wrobel@emory.edu}
}

Other contributors:
\itemize{
  \item Junting Ren \email{junting.ren.stat@gmail.com} [contributor]
}

}
\keyword{internal}
