% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valscore.R
\name{valscore}
\alias{valscore}
\title{Pré-validation d'un score
Calcule le tableau du risque et des effectifs par niveau, avec son histogramme, courbe ROC avec l'AUC.}
\usage{
valscore(sc, varx, titre = "Score")
}
\arguments{
\item{sc}{: score à tester}

\item{varx}{: variable d'outcome}

\item{titre}{: titre du graphique}
}
\value{
Tableau, histogramme, courbe ROC
}
\description{
Pré-validation d'un score
Calcule le tableau du risque et des effectifs par niveau, avec son histogramme, courbe ROC avec l'AUC.
}
\examples{
a = 2
}
